+++ bigframes/series.py
@@ -43,7 +43,7 @@ import bigframes.core.indexes as indexes
 import bigframes.core.ordering as order
 import bigframes.core.scalar as scalars
 import bigframes.core.utils as utils
-import bigframes.core.validate as validations
+import bigframes.core.validations as validations
 import bigframes.core.window
 import bigframes.core.window_spec
 import bigframes.dataframe
